almostdone

+++ studentlist.php
@@ -1,4 +1,5 @@
 <?php 
+
 include('DBheader.php');
 require_once('database.php');
 ?>
@@ -15,10 +16,9 @@ $result = mysqli_query($connection,$info);
     <div class="p-5">
             <h1 class="display-4 text-center text-light py-2 rgb mb-5 mx-auto">Students list</h1>
         
-            <table class="table table-light table-striped text-info">
+            <table class="table table-responsive-sm table-sm  table-light table-striped table-hover text-info pl-2">
                 <thead>
                     <tr class="font">
-                    <th scope="col"class="text-info">Id</th>
                     <th scope="col">First Name</th>
                     <th scope="col">Last Name</th>
                     <th scope="col">Phone Number</th>
@@ -32,7 +32,6 @@ $result = mysqli_query($connection,$info);
                 while ($lop=mysqli_fetch_assoc($result)){
                 ?>
                     <tr class="text-left text-info">
-                    <th scope="row"class="text-info"><?php echo $lop['id'];?></th>
                     <td><?php echo $lop['first_name'];?></td>
                     <td><?php echo $lop['last_name'];?></td>
                     <td><?php echo $lop['mobile'];?></td>
@@ -40,9 +39,11 @@ $result = mysqli_query($connection,$info);
                     <td><?php echo $lop['email'];?></td>
                     <td><?php echo $lop['student_pass'];?></td>
     
-                    <td class="text-info ">
+                    <td class="text-info float-right">
+                   <?php if(isset($_SESSION['username'])){?>
                     <a href="studentedit.php?id=<?php echo $lop['id'];?>" class="btn xs btn-warning">Edit</a> ||
                     <a href="studentdelete.php?id=<?php echo $lop['id']?>" class="btn xs btn-danger text-light">Delete</a> ||
+                    <?php }?>
                     <a href="studentshow.php?id=<?php echo $lop['id'];?>" class="btn xs btn-success">Show</a>
                     </td>
                     </tr>
